+++ src/main/java/com/troyboot/java/system/dao/SysPermissionMapper.java
@@ -3,13 +3,13 @@ package com.troyboot.java.system.dao;
 import com.troyboot.java.system.po.SysPermission;
 
 public interface SysPermissionMapper {
-    int deleteByPrimaryKey(Long id);
+    int deleteByPrimaryKey(Integer id);
 
     int insert(SysPermission record);
 
     int insertSelective(SysPermission record);
 
-    SysPermission selectByPrimaryKey(Long id);
+    SysPermission selectByPrimaryKey(Integer id);
 
     int updateByPrimaryKeySelective(SysPermission record);
