+++ src/main/java/com/troyboot/java/system/dao/SysRoleMapper.java
@@ -3,13 +3,13 @@ package com.troyboot.java.system.dao;
 import com.troyboot.java.system.po.SysRole;
 
 public interface SysRoleMapper {
-    int deleteByPrimaryKey(Long id);
+    int deleteByPrimaryKey(Integer id);
 
     int insert(SysRole record);
 
     int insertSelective(SysRole record);
 
-    SysRole selectByPrimaryKey(Long id);
+    SysRole selectByPrimaryKey(Integer id);
 
     int updateByPrimaryKeySelective(SysRole record);
