+++ src/main/java/com/troyboot/java/system/dao/SysUserMapper.java
@@ -3,13 +3,13 @@ package com.troyboot.java.system.dao;
 import com.troyboot.java.system.po.SysUser;
 
 public interface SysUserMapper {
-    int deleteByPrimaryKey(Long id);
+    int deleteByPrimaryKey(Integer id);
 
     int insert(SysUser record);
 
     int insertSelective(SysUser record);
 
-    SysUser selectByPrimaryKey(Long id);
+    SysUser selectByPrimaryKey(Integer id);
 
     int updateByPrimaryKeySelective(SysUser record);
